backend/requirements.txt
# API Framework
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
email-validator==2.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
httpx==0.24.1
python-dotenv==1.0.0
aiohttp==3.9.1

# Web Scraping & Content Processing
beautifulsoup4==4.12.2
aiocron==1.8
playwright==1.41.0
requests==2.31.0
pydub==0.25.1  # For audio file manipulation
pytube==15.0.0  # For YouTube video downloading
moviepy==1.0.3  # For video processing and audio extraction
yt-dlp==2023.11.16  # Advanced YouTube downloader (alternative to pytube)

# Azure Services
azure-identity==1.13.0
azure-search-documents==11.4.0  # Stable version for Azure Search
azure-ai-formrecognizer==3.2.0
azure-ai-textanalytics==5.2.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-cognitiveservices-speech==1.32.1  # For speech-to-text processing
msrest==0.7.1

# OpenAI with Azure
openai==1.10.0  # Specifically using 1.10.0 version

# LangChain Integration - Compatible set of versions
langchain==0.1.20
langchain-core==0.1.53
langchain-openai==0.0.2
langchain-community==0.0.38
langchain-text-splitters==0.0.2
tiktoken==0.5.2  # For token counting

# Testing
pytest==7.3.1

# Utilities
tabulate==0.9.0  # For formatted table output in scripts
python-dateutil==2.8.2  # For date parsing